+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Cert.hs
@@ -54,10 +54,8 @@ import Cardano.Ledger.Conway.TxCert (
  )
 import Cardano.Ledger.Shelley.API (
   CertState (..),
-  DState,
   PState (..),
   PoolEnv (PoolEnv),
-  VState,
  )
 import Cardano.Ledger.Shelley.Rules (PoolEvent, ShelleyPOOL, ShelleyPoolPredFailure)
 import Control.DeepSeq (NFData)
@@ -174,9 +172,9 @@ instance
 instance
   forall era.
   ( Era era
-  , State (EraRule "DELEG" era) ~ DState era
+  , State (EraRule "DELEG" era) ~ CertState era
   , State (EraRule "POOL" era) ~ PState era
-  , State (EraRule "GOVCERT" era) ~ VState era
+  , State (EraRule "GOVCERT" era) ~ CertState era
   , Environment (EraRule "DELEG" era) ~ ConwayDelegEnv era
   , Environment (EraRule "POOL" era) ~ PoolEnv era
   , Environment (EraRule "GOVCERT" era) ~ ConwayGovCertEnv era
@@ -201,9 +199,9 @@ instance
 
 certTransition ::
   forall era.
-  ( State (EraRule "DELEG" era) ~ DState era
+  ( State (EraRule "DELEG" era) ~ CertState era
   , State (EraRule "POOL" era) ~ PState era
-  , State (EraRule "GOVCERT" era) ~ VState era
+  , State (EraRule "GOVCERT" era) ~ CertState era
   , Environment (EraRule "DELEG" era) ~ ConwayDelegEnv era
   , Environment (EraRule "POOL" era) ~ PoolEnv era
   , Environment (EraRule "GOVCERT" era) ~ ConwayGovCertEnv era
@@ -217,22 +215,19 @@ certTransition ::
   ) =>
   TransitionRule (ConwayCERT era)
 certTransition = do
-  TRC (CertEnv slot pp currentEpoch committee committeeProposals, cState, c) <- judgmentContext
+  TRC (CertEnv slot pp currentEpoch committee committeeProposals, certState, c) <- judgmentContext
   let
-    CertState {certDState, certPState, certVState} = cState
+    CertState {certPState} = certState
     pools = psStakePoolParams certPState
   case c of
     ConwayTxCertDeleg delegCert -> do
-      newDState <- trans @(EraRule "DELEG" era) $ TRC (ConwayDelegEnv pp pools, certDState, delegCert)
-      pure $ cState {certDState = newDState}
+      trans @(EraRule "DELEG" era) $ TRC (ConwayDelegEnv pp pools, certState, delegCert)
     ConwayTxCertPool poolCert -> do
       newPState <- trans @(EraRule "POOL" era) $ TRC (PoolEnv slot pp, certPState, poolCert)
-      pure $ cState {certPState = newPState}
+      pure $ certState {certPState = newPState}
     ConwayTxCertGov govCert -> do
-      newVState <-
-        trans @(EraRule "GOVCERT" era) $
-          TRC (ConwayGovCertEnv pp currentEpoch committee committeeProposals, certVState, govCert)
-      pure $ cState {certVState = newVState}
+      trans @(EraRule "GOVCERT" era) $
+        TRC (ConwayGovCertEnv pp currentEpoch committee committeeProposals, certState, govCert)
 
 instance
   ( Era era