+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/LedgerTypes/Specs.hs
@@ -353,7 +353,7 @@ instance IsConwayUniv fn => NumLike fn EpochNo
 
 drepStateSpec :: (IsConwayUniv fn, Crypto c) => Term fn EpochNo -> Specification fn (DRepState c)
 drepStateSpec epoch = constrained $ \ [var|drepstate|] ->
-  match drepstate $ \ [var|expiry|] _anchor [var|drepDdeposit|] ->
+  match drepstate $ \ [var|expiry|] _anchor [var|drepDdeposit|] _delegs ->
     [ assertExplain (pure "epoch of expiration must follow the current epoch") $ epoch <=. expiry
     , assertExplain (pure "no deposit is 0") $ lit (Coin 0) <=. drepDdeposit
     ]
@@ -374,14 +374,25 @@ dstateSpec ::
   LedgerEra era fn =>
   Term fn AccountState ->
   Term fn (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era))) ->
+  Term fn (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era))) ->
   Specification fn (DState era)
-dstateSpec acct poolreg = constrained $ \ [var| ds |] ->
+dstateSpec acct poolreg dreps = constrained $ \ [var| ds |] ->
   match ds $ \ [var|umap|] [var|futureGenDelegs|] [var|genDelegs|] [var|irewards|] ->
-    match umap $ \ [var|rdMap|] [var|ptrmap|] [var|sPoolMap|] _dRepMap ->
+    match umap $ \ [var|rdMap|] [var|ptrmap|] [var|sPoolMap|] [var|dRepMap|] ->
       [ genHint 5 sPoolMap
       , assertExplain (pure "The delegations delegate to actual pools") $
           forAll (rng_ sPoolMap) (\ [var|keyhash|] -> member_ keyhash (dom_ poolreg))
       , assertExplain (pure "dom sPoolMap is a subset of dom rdMap") $ dom_ sPoolMap `subset_` dom_ rdMap
+      , -- NOTE: Consider if this assertion (and the `dependsOn` check below it) can be removed.
+        -- Commit `21215b03a - Add delegations field to the DRep state` added a TODO
+        -- to add a constraint that delegs are in the UMap. The below does that but it wasn't
+        -- the cause for the conformance test failures.
+        forAll' dreps $
+          \_ dRepState -> match dRepState $ \_ _ _ delegs ->
+            assertExplain
+              (pure "Delegs are present in the UMap")
+              (forAll delegs (\ [var|drep|] -> drep `member_` dom_ dRepMap))
+      , dreps `dependsOn` dRepMap
       , -- reify here, forces us to solve for ptrap, before sovling for rdMap
         whenTrue (hasPtrs (Proxy @era)) (reify ptrmap id (\ [var|pm|] -> domEqualRng pm rdMap))
       , whenTrue (not_ (hasPtrs (Proxy @era))) (assert $ ptrmap ==. lit Map.empty)
@@ -434,7 +445,17 @@ certStateSpec acct epoch = constrained $ \ [var|certState|] ->
   match certState $ \ [var|vState|] [var|pState|] [var|dState|] ->
     [ satisfies vState (vstateSpec epoch)
     , satisfies pState (pstateSpec epoch)
-    , reify pState psStakePoolParams (\ [var|poolreg|] -> satisfies dState (dstateSpec acct poolreg))
+    , reify
+        pState
+        psStakePoolParams
+        ( \ [var|poolreg|] ->
+            reify
+              vState
+              vsDReps
+              ( \ [var|dreps|] ->
+                  satisfies dState (dstateSpec acct poolreg dreps)
+              )
+        )
     ]
 
 -- ==============================================================