+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/STS.hs
@@ -181,7 +181,7 @@ prop_DELEG :: Property
 prop_DELEG =
   stsPropertyV2 @"DELEG" @ConwayFn
     delegEnvSpec
-    (\_env -> dStateSpec)
+    (\_env -> certStateSpec)
     delegCertSpec
     $ \_env _st _sig _st' -> True
 
@@ -197,7 +197,7 @@ prop_GOVCERT :: Property
 prop_GOVCERT =
   stsPropertyV2 @"GOVCERT" @ConwayFn
     govCertEnvSpec
-    (\_env -> vStateSpec)
+    (\_env -> certStateSpec)
     (\env st -> govCertSpec env st)
     $ \_env _st _sig _st' -> True